Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/svg/path/property/getComputedStyle-expected.txt


FAIL getComputedStyle assert_equals: expected (string) "none" but got (undefined) undefined
